Refactor orientation change handler to separate file
