+++ src/commons/hooks/useOrientationChangeHandler.ts
@@ -1,0 +1,47 @@
+import { useEffect, useRef } from 'react';
+
+/**
+ * The following effect prevents the mobile browser interface from hiding on scroll by setting the
+ * application height to the window's innerHeight, even after orientation changes. This ensures that
+ * the app UI does not break due to the hiding of the browser interface when the user is not on the PWA.
+ *
+ * Note: When the soft keyboard is up on Android devices, the viewport height decreases and triggers
+ * the 'resize' event. The conditional in orientationChangeHandler checks specifically for this, and
+ * does not update the application height when the Android keyboard triggers the resize event. IOS
+ * devices are not affected.
+ */
+export function useOrientationChangeHandler() {
+  // Used in the mobile/PWA experience (e.g. separate handling of orientation changes on Andriod & iOS due to unique browser behaviours)
+  const isMobile = /iPhone|iPad|Android/.test(navigator.userAgent);
+  const isPWA = window.matchMedia('(display-mode: standalone)').matches; // Checks if user is accessing from the PWA
+  const browserDimensions = useRef({ height: 0, width: 0 });
+
+  useEffect(() => {
+    const orientationChangeHandler = () => {
+      if (
+        !(
+          window.innerHeight < browserDimensions.current.height &&
+          window.innerWidth === browserDimensions.current.width
+        )
+      ) {
+        // If it is not an Android soft keyboard triggering the resize event, update the application height.
+        document.documentElement.style.setProperty(
+          '--application-height',
+          window.innerHeight + 'px',
+        );
+      }
+      browserDimensions.current = { height: window.innerHeight, width: window.innerWidth };
+    };
+
+    if (!isPWA && isMobile) {
+      orientationChangeHandler();
+      window.addEventListener('resize', orientationChangeHandler);
+    }
+
+    return () => {
+      if (!isPWA && isMobile) {
+        window.removeEventListener('resize', orientationChangeHandler);
+      }
+    };
+  }, [isPWA, isMobile]);
+}
